Friends_Back/requirements.txt
asgiref==3.5.2
boto3==1.24.21
botocore==1.27.21
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.0
Django==3.2.13
django-cors-headers==3.13.0
django-storages==1.12.3
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
google-api-core==2.8.2
google-api-python-client==2.52.0
google-auth==2.9.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.2
googleapis-common-protos==1.56.3
httplib2==0.20.4
idna==3.3
jmespath==1.0.1
oauthlib==3.2.0
protobuf==4.21.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==2.4.0
PyMySQL==1.0.2
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.8
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.2.0
uritemplate==4.1.1
urllib3==1.26.9
